feat: legend
