+++ apps/web/components/dashboard/common/TileGrid.tsx
@@ -4,16 +4,10 @@ import { useContentMoreMenu } from "@/hooks/dashboard/ContentHooks";
 import type { Layer } from "@/lib/validations/layer";
 import type { Project } from "@/lib/validations/project";
 import type { ContentActions } from "@/types/common";
-import {
-  Box,
-  Grid,
-  Skeleton,
-  Stack,
-  Typography,
-  useTheme,
-} from "@mui/material";
-import { ICON_NAME, Icon } from "@p4b/ui/components/Icon";
+import { Box, Grid, Skeleton } from "@mui/material";
+import { ICON_NAME } from "@p4b/ui/components/Icon";
 import { useTranslation } from "@/i18n/client";
+import EmptySection from "@/components/common/EmptySection";
 
 interface TileGridProps {
   view: "list" | "grid";
@@ -27,7 +21,6 @@ interface TileGridProps {
 
 const TileGrid = (props: TileGridProps) => {
   const { items, isLoading } = props;
-  const theme = useTheme();
   const { t } = useTranslation("dashboard");
   const listProps = {
     xs: 12,
@@ -71,31 +64,16 @@ const TileGrid = (props: TileGridProps) => {
         <Grid container spacing={props.view === "list" ? 0 : 5}>
           {!isLoading && items?.length === 0 && (
             <Grid item xs={12}>
-              <Stack
-                direction="column"
-                spacing={4}
-                sx={{
-                  mt: 10,
-                  display: "flex",
-                  justifyContent: "center",
-                  alignItems: "center",
-                  height: "100%",
-                }}
-              >
-                <Icon
-                  iconName={
-                    props.type === "project"
-                      ? ICON_NAME.MAP
-                      : ICON_NAME.DATABASE
-                  }
-                  htmlColor={theme.palette.text.secondary}
-                />
-                <Typography variant="h6" color={theme.palette.text.secondary}>
-                  {props.type === "project"
+              <EmptySection
+                label={
+                  props.type === "project"
                     ? t("projects.no_projects_found")
-                    : t("projects.no_datasets_found")}
-                </Typography>
-              </Stack>
+                    : t("projects.no_datasets_found")
+                }
+                icon={
+                  props.type === "project" ? ICON_NAME.MAP : ICON_NAME.DATABASE
+                }
+              />
             </Grid>
           )}
           {(isLoading ? Array.from(new Array(4)) : items ?? []).map(
